Move debugPage to /pages
